Change while-loop for reading input
Both the client and the server now read
an individual element from the input stream
and check if it does not equal -1. This allows
the program to continue until there is no more
input in the stream. Removed Printwriter in the
server. All tests are passing for the server,
client tests are hanging.